REFACTOR: corrected the package directions
